staging: fsl-mc/dpio: Change bpid type to u16
In all QBMan registers, the buffer pool id field is two bytes
long. Modify both the high level DPIO APIs and the low level
qbman ones to reflect this, in order to avoid (potentially buggy)
implicit downcasts.

+++ drivers/staging/fsl-mc/bus/dpio/qbman-portal.c
@@ -1127,7 +1127,7 @@ struct qbman_bp_query_desc {
 	u8 reserved2[60];
 };
 
-int qbman_bp_query(struct qbman_swp *s, u32 bpid,
+int qbman_bp_query(struct qbman_swp *s, u16 bpid,
 		   struct qbman_bp_query_rslt *r)
 {
 	struct qbman_bp_query_desc *p;
